feat: hide row when no visible elements
If no elements in the row are visible due to conditions being false, the row will not render an empty div.
Additionally refactors the logic into a SchemaRow intermediate component, this was needed due to not being able to modify the loop. The `parsedSchema` computed MUST be used or else a refactor of the plugins would be required, so the logic had to be enclosed into a new component.

closes: #208