fix: lookup mapComp and mapProp race condition
When using `lookupSubSchemas` and mapProps/mapComponents at the same time a race condition happened.
Due to the lookupSubSchemas parsing function being called at ParsedSchema.js level, it silently "skipped" it if the component property was not found.
This required refactoring the approach into Lookup providing the actual function that needs to be executed when this happens, so that we can eagerly call it on ParsedSchema.js
for scenarios where `lookupSubSchemas` is used _without_ the LookupPlugin being added to the plugin stack AND scenarios where the plugin is also added.